requirements.txt
huggingface-hub>=0.34.0
groq>=0.11.0
transformers>=4.55.0
torch>=2.6.0
sentence-transformers>=3.0.0
chromadb>=1.5.0
streamlit>=1.50.0
requests>=2.32.0
beautifulsoup4>=4.13.0
pypdf>=6.0.0
python-dotenv>=1.2.0
SQLAlchemy>=2.0.0
rank-bm25>=0.2.2
